AiCHATBOT/requirements.txt
flask==3.0.0
openai==1.3.0
ollama==0.1.7
python-dotenv==1.0.0
flask-cors==4.0.0
numpy==1.24.3
werkzeug==3.0.1

